light

+++ faiss_cli.py
@@ -18,7 +18,7 @@ def IndexDB(file, d):
 		db.append(l)
 
 	index = faiss.IndexFlatL2(d)  		# build the index
-	db = np.array(db,dtype=float)
+	db = np.array(db,dtype=float32)
 	print(db.shape)
 	index.add(db) # add vectors to the index
 	print(index.ntotal)
